refactor: Json変換不能時のエラー処理をAdviceに移動、システムで一元化。
