--- src/main/java/com/github/kazuhito_m/mysample/presentation/ControllerAdviceConfiguration.java
@@ -1,13 +1,0 @@
-package com.github.kazuhito_m.mysample.presentation;
-
-import org.springframework.validation.DataBinder;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.InitBinder;
-
-@ControllerAdvice
-public class ControllerAdviceConfiguration {
-    @InitBinder
-    private void initDirectFieldAccess(DataBinder dataBinder) {
-        dataBinder.initDirectFieldAccess();
-    }
-}
