+++ src/main/java/com/github/kazuhito_m/mysample/presentation/CustomRestControllerAdvice.java
@@ -1,0 +1,30 @@
+package com.github.kazuhito_m.mysample.presentation;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.validation.DataBinder;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.InitBinder;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class CustomRestControllerAdvice {
+    @InitBinder
+    private void initDirectFieldAccess(DataBinder dataBinder) {
+        dataBinder.initDirectFieldAccess();
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    private Invalidate handleException(MethodArgumentNotValidException e) {
+        return new Invalidate(e, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(HttpMessageNotReadableException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    private Invalidate handleException(HttpMessageNotReadableException e) {
+        return new Invalidate(e, HttpStatus.BAD_REQUEST);
+    }
+}
